Refactoring Changes
