+++ client/src/Pages/SoftwaresAndPackages/WebApplications.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StyledSoftware } from './StyledSoftwaresAndPackages';
+import { SoftwareAndPackageStyles } from './SoftwareAndPackageStyles';
 import getRandomElementsFromArray from '../../utils/getRandomElementsFromArray';
 
 const WEBAPPLICATION_DISPLAY_COUNT = 4;
@@ -86,7 +86,7 @@ const subsetOfWebApplications = getRandomElementsFromArray(
 
 function WebApplications() {
 	return (
-		<StyledSoftware>
+		<SoftwareAndPackageStyles>
 			<div className="component-heading">
 				<h1>
 					Softwares
@@ -118,7 +118,7 @@ function WebApplications() {
 					))
 				}
 			</div>
-		</StyledSoftware>
+		</SoftwareAndPackageStyles>
 	);
 }
